+++ app/src/main/java/com/project/symptoms/db/model/SelectedCategoryOptionModel.java
@@ -2,18 +2,18 @@ package com.project.symptoms.db.model;
 
 public class SelectedCategoryOptionModel {
 
-    private int symptomId, categoryOptionId;
+    private long symptomId, categoryOptionId;
 
-    public SelectedCategoryOptionModel(int symptomId, int categoryOptionId) {
+    public SelectedCategoryOptionModel(long symptomId, long categoryOptionId) {
         this.symptomId = symptomId;
         this.categoryOptionId = categoryOptionId;
     }
 
-    public int getSymptomId() { return symptomId; }
+    public long getSymptomId() { return symptomId; }
 
     public void setSymptomId(int symptomId) { this.symptomId = symptomId; }
 
-    public int getCategoryOptionId() { return categoryOptionId; }
+    public long getCategoryOptionId() { return categoryOptionId; }
 
     public void setCategoryOptionId(int categoryOptionId) { this.categoryOptionId = categoryOptionId; }
 }
